+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the ADTool will be documented in this file.
 - added copy/paste of values in Domain View
 - fixed not updating node size after checking "Computed Values" checkbox
 - added copy/paste of values in Valuations View
+- added command line options
 
 ## [2.0.1] - 2016-03-01
 - fixed bug with wrong Sand Tree terms